[wip #39] manage active ZMQ subscriptions per accessor
This hopes to fix a potential race condition during recovery, since DOOCS renews the subscription simultaneously with the backend (previously) re-subscribing the property.

Additionally, it fixes wrong behaviour if multiple instances of the DOOCS backend access the same properties, in which case the subscription status was shared across the accessors from the different backends.