chore: extend TransferGroup tests (cf. #14823)
The failure was actually unrelated to the DoocsBackend and the tests are
a bit redundant here (i.e. testing mainly TransferElement base class
behaviour). It should be sufficient to keep the two, newly added
"mayReplaceOther" checks as well as the three
getHardwareAccessingElements() tests. The other tests are basically an
integration test, which would be unnecessary if we trust the
TransferGroup and TransferElement base classes to follow the spec. We
keep it for now, since there is apparently a gab in the DeviceAccess
tests.